Move XR State Manager under menu
